+++ tgsdk/entities/files/video_note.py
@@ -33,10 +33,10 @@ class VideoNote(TelegramEntity):
 		file_unique_id: str,
 		length: int,
 		duration: int,
-		thumb: PhotoSize = None,
-		file_size: int = None,
+		thumb: Optional[PhotoSize] = None,
+		file_size: Optional[int] = None,
 
-		bot: "Bot" = None
+		bot: Optional["Bot"] = None
 	):
 		self.file_id = file_id
 		self.file_unique_id = file_unique_id